requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.9.2
python-dotenv==1.0.1
groq==0.9.0
pymongo==4.8.0
chromadb
streamlit
requests==2.32.3
httpx==0.27.0
python-multipart==0.0.9